Update the structure of the COOL-core, add the cool.functionality and the cool.result
