+++ cool-core/src/main/java/com/nus/cool/model/CoolCohortEngine.java
@@ -1,15 +1,16 @@
 package com.nus.cool.model;
 
-import com.nus.cool.core.cohort.CohortUserSection;
-import com.nus.cool.core.cohort.ExtendedCohortQuery;
-import com.nus.cool.core.cohort.ExtendedCohortSelection;
+import com.nus.cool.core.cohort.*;
 import com.nus.cool.core.io.compression.Compressor;
 import com.nus.cool.core.io.compression.Histogram;
 import com.nus.cool.core.io.compression.ZIntBitCompressor;
 import com.nus.cool.core.io.readstore.*;
 import com.nus.cool.core.io.storevector.InputVector;
+import com.nus.cool.core.schema.DataType;
+import com.nus.cool.core.schema.FieldSchema;
 import com.nus.cool.core.schema.FieldType;
 import com.nus.cool.core.schema.TableSchema;
+import com.nus.cool.result.ExtendedResultTuple;
 
 import java.io.DataOutputStream;
 import java.io.File;
@@ -18,6 +19,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
+import java.util.Map;
 
 public class CoolCohortEngine {
 
@@ -96,4 +98,109 @@ public class CoolCohortEngine {
         stream.close();
     }
 
+    public static List<ExtendedResultTuple> processCohortResult(ExtendedCohortQuery query,
+                                                                TableSchema tableSchema,
+                                                                MetaChunkRS metaChunk,
+                                                                ExtendedCohortSelection sigma,
+                                                                Map<ExtendedCohort, Map<Integer, List<Double>>> resultMap){
+        List<ExtendedResultTuple> resultSet = new ArrayList<>();
+        for (Map.Entry<ExtendedCohort, Map<Integer, List<Double>>> entry:resultMap.entrySet()) {
+            ExtendedCohort key = entry.getKey();
+
+            // update dimension names
+            List<BirthSequence.BirthEvent> events = query.getBirthSequence().getBirthEvents();
+            int cnt = 0;
+            for (int idx = 0; idx < events.size(); ++idx) {
+                BirthSequence.BirthEvent be = events.get(idx);
+                for (BirthSequence.CohortField cf : be.getCohortFields()) {
+                    String fieldName = cf.getField();
+                    int fieldID = tableSchema.getFieldID(cf.getField());
+                    FieldSchema schema = tableSchema.getField(fieldID);
+                    if (schema.getDataType() == DataType.String) {
+                        key.addDimensionName(metaChunk.getMetaField(fieldName).getString(key.getDimensions().get(cnt)));
+                    } else {
+                        MetaFieldRS filed = metaChunk.getMetaField(fieldName);
+                        StringBuilder builder = new StringBuilder();
+                        builder.append("(");
+                        int level = key.getDimensions().get(cnt);
+                        if (cf.getMinLevel() == level) {
+                            builder.append(Integer.toString(filed.getMinValue())+", ");
+                        } else {
+                            if (cf.isLogScale()) {
+                                builder.append((new Double(Math.pow(cf.getScale(), level - 1))).longValue());
+                            } else {
+                                builder.append((new Double(cf.getScale() * (level - 1))).longValue());
+                            }
+                            builder.append(", ");
+                        }
+
+                        if (cf.getMinLevel() + cf.getNumLevel() == level) {
+                            builder.append(Integer.toString(filed.getMaxValue())+")");
+                        } else {
+                            if (cf.isLogScale()) {
+                                builder.append((new Double(Math.pow(cf.getScale(), level))).longValue());
+                            } else {
+                                builder.append((new Double(cf.getScale() * level)).longValue());
+                            }
+                            builder.append("]");
+                        }
+                        key.addDimensionName(builder.toString());
+                    }
+                    ++cnt;
+                }
+            }
+
+            for (Map.Entry<Integer, List<Double>> ageMetric : entry.getValue().entrySet()) {
+                int age = ageMetric.getKey();
+                if (!sigma.getAgeFieldFilter().accept(age))
+                    continue;
+                double measure = ageMetric.getValue().get(0);
+                double min = 0.0;
+                double max = 0.0;
+                double sum = 0.0;
+                double num = 0.0;
+                if (ageMetric.getValue().size() >= 5) {
+                    min = ageMetric.getValue().get(1);
+                    max = ageMetric.getValue().get(2);
+                    sum = ageMetric.getValue().get(3);
+                    num = ageMetric.getValue().get(4);
+                }
+                if (num==0) {
+                    min = 0.0;
+                    max = 0.0;
+                }
+                resultSet.add(new ExtendedResultTuple(key.toString(), age, measure, min, max, sum, num));
+            }
+        }
+        return resultSet;
+    }
+
+    public static List<ExtendedResultTuple> performCohortQuery(CubeRS cube, InputVector users, ExtendedCohortQuery query) {
+        List<CubletRS> cublets = cube.getCublets();
+        TableSchema tableSchema = cube.getTableSchema();
+        List<ExtendedResultTuple> resultSet = new ArrayList<>();
+
+        for (CubletRS cubletRS : cublets) {
+            MetaChunkRS metaChunk = cubletRS.getMetaChunk();
+            ExtendedCohortSelection sigma = new ExtendedCohortSelection();
+            ExtendedCohortAggregation gamma = new ExtendedCohortAggregation(sigma);
+
+            // Initialize the birth events and trigger times
+            gamma.init(tableSchema, users, query);
+            // Check the validity of fields (e.g., the existences of field names and their values)
+            gamma.process(metaChunk);
+            if (sigma.isUserActiveCublet()) {
+                List<ChunkRS> dataChunks = cubletRS.getDataChunks();
+                for (ChunkRS dataChunk : dataChunks) {
+                    gamma.process(dataChunk);
+                }
+            }
+
+            Object results = gamma.getCubletResults();
+
+            resultSet.addAll(processCohortResult(query,tableSchema,metaChunk,sigma,(Map<ExtendedCohort, Map<Integer, List<Double>>>) results));
+        }
+
+        return resultSet;
+    }
 }
