+++ cool-core/src/main/java/com/nus/cool/model/CoolLoader.java
@@ -79,21 +79,5 @@ public class CoolLoader {
         Files.copy(schemaFile, new File(outputCubeVersionDir, "table.yaml"));
     }
 
-    public static void main(String[] args) {
-        String cube = args[0];
-        String schemaFileName = args[1];
-        String dimensionFileName = args[2];
-        String dataFileName = args[3];
-        String cubeRepo = args[4];
 
-        try {
-            DataLoaderConfig config = new CsvDataLoaderConfig();
-            CoolLoader coolLoader = new CoolLoader(config);
-            coolLoader.load(cube,schemaFileName,dimensionFileName,dataFileName,cubeRepo);
-        } catch (IOException e){
-            System.out.println(e);
-        }
-        System.out.println("Cube " + cube + " is loaded successfully from the CSV format data.");
-
-    }
 }
