+++ cool-extensions/avro-extensions/src/main/java/com/nus/cool/extension/model/AvroLoader.java
@@ -9,8 +9,7 @@ import java.io.IOException;
 
 public class AvroLoader {
     /**
-     * Please list the sogamo dataset files, because we are generating
-     *  a sample avro file for testing according to that dataset,
+     * Please list the necessary dataset files for the COOL system to load into a new cube.
      * @param args there are five arguments. List in input order
      *  (1) output cube name: to be specified when loading from the repository
      *  (2) table.yaml (3) dimension.csv (4) data.avro (5) output cube repository
