+++ pom.xml
@@ -15,8 +15,8 @@
 
     <modules>
         <module>cool-core</module>
-        <module>cool-extensions</module>
-        <module>cool-queryserver</module>
+<!--        <module>cool-extensions</module>-->
+<!--        <module>cool-queryserver</module>-->
     </modules>
